Change context size and reasoning budget types from int64/uint64 to int32 (#487)
* fix: change context size and reasoning budget types from uint64 to int32

* fix: change context size and reasoning budget types from int64 to int32

+++ cmd/mdltool/main.go
@@ -321,7 +321,7 @@ func cmdPackage(args []string) int {
 
 	if contextSize > 0 {
 		fmt.Println("Setting context size:", contextSize)
-		b = b.WithContextSize(contextSize)
+		b = b.WithContextSize(int32(contextSize))
 	}
 
 	if mmproj != "" {
